Clean up example-1
